requirements.txt
certifi==2020.12.5
chardet==4.0.0
ciso8601==2.2.0
enum-compat==0.0.3
idna==2.10
influxdb-client==1.25.0
msgpack==1.0.2
pexpect==4.8.0
pip==21.0.1
ptyprocess==0.7.0
pygatt==4.0.5
pyserial==3.5
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
Rx==3.1.1
setuptools==49.2.1
six==1.15.0
tqdm==4.56.0
urllib3==1.26.3
